[Examples] Fix species indexing in diamond_cvd.py
The default ordering of phases when constructing InterfaceKinetics
objects changed in the YAML format to put the reacting phase first.

Since the species names are unique across all phases, the easiest way
to look up their index in the Kinetics object is by name, which avoids
needing to think about the phase ordering.