This commit is the rough implementation of the adaptive preconditioner.
It adds PreconditionerBase and derives AdaptivePreconditioner from this
base class. PreconditionerBase is intended to be an abstract class so
other types of preconditioners can be extended. Severla tests for these
classes are also added. AdaptivePreconditioner is a specific type of
preconditioning based on a tolerance. Other functions were added as was
necessary to create a robust and extensible preconditioner that
interacts well with the zeroD reactor system. A simple python interface
was also created. Currently, it integrates but there are some comparison
that need to be made to ensure that it everything is working properly.