This commit adds the MoleReactor and IdealGasMoleReactor classes.
The former is intended to be the basis for the mole based reactor systems similar to how Reactor is the basis for the current subclasses. It currently does not integrate but allows common methods to be used by subclasses. The latter is an ideal gas constant volume mole reactor designed for use with the preconditioner. It currently fails the wall velocity test but passes all other tests. This commit also adds testing for these classes and preconditioning.