This commit refactors the preconditioner interfaces
This commit updates the preconditioners interface with reactors, cvodes, and the reactor network. It moves the preconditioning to a reactor basis similar to evalEqs where the preconditioner is supplied to the reactor and not the other way around. It separates the preconditioning type from problem type to provide a cleaner interface to adjust problem and preconditioning types separately. It updates the python interface according to these changes. Finally, stylistic, comment, and formatting changes were made.